+++ libraries/source/fcollada/src/FCollada/FMath/FMVector3.cpp
@@ -64,7 +64,7 @@ FMVector4 FMVector4::FromHSVColor(float hue, float saturation, float value)
 	if (!IsEquivalent(saturation, 0.0f)) 
 	{
 		hue *= 6.0f;						// sector 0 to 5
-		float sector = floor(hue);
+		float sector = (float)floor(hue);
 		float f = hue - sector;				// factorial part of h
 		float p = value * (1.0f - saturation);
 		float q = value * (1.0f - saturation * f);
